split server/commands/mod.rs
